classes

+++ Classes/Calculator.cs
@@ -14,7 +14,7 @@ namespace Classes
             int sum = numOne + numTwo;
             return sum;
         }
-        public int Add(double numOne, double numTwo)
+        public double Add(double numOne, double numTwo)
         {
             double sum = numOne + numTwo;
             return sum;
